+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - `Repeat@FlowType_K` を追加
 - プロパティメニューに`プロパティ名をコピー`と`参照式をコピー`を追加
 - `Motion@FlowType_K` に `Offset` を追加
+- `Motion@FlowType_K` と `Blink@FlowType_K` に文字置換機能を追加
 - `Kerning@FlowType_K` に `Limit Effects Below` を追加
 - `Blink@FlowType_K` に `Based On` を追加
 - `Island@FlowType_K` で `Time Offset::Order` によって個別オブジェクトのインデックス付与順が変わるように変更